Add files via upload
It is a dice game that determines the winner based on the total score you get after you rolled your device consecutively. But if you get a 1, your total score drops back down to a zero. And displays the final answer when the input to to the question "would you like to roll the dice" is given as n.